pyproject.toml
[project]
name = "eidola"
version = "0.1.0"
description = "Multi-agent Instagram automation system using Google ADK and FIRERPA"
requires-python = ">=3.11"

dependencies = [
    "google-adk",
    "google-cloud-aiplatform",
    "google-genai",
    "lamda",
    "motor",
    "pymongo",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "httpx",
    "pyyaml",
    "pyotp",
    "aiogram",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eidola"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
